Add files via upload
This directory contains the data which were used for the prediction of our model. The first dataset is referenced by PICK108, it is composed of 108 complete bacterial genomic sequences and their annotations. The number of positive and negative GIs for this data set was 3868 and 679 respectively. The second set of data is referenced by CF15. It is composed of 15 complete bacterial genomic sequences and their annotations. The number of positive and negative GIs for this data set were 6070 and 5833 respectively. And the third dataset is referenced by RGP104, it is composed of 104 complete bacterial genomic sequences and their annotations. The positive and negative genomic island number for this dataset is 1846 and 3267 respectively.
the other three files contain the results of the experimentation of our three datasets, each file contains 3 excel sheets, the first presents the evaluation of our different feature extraction on our model, the second sheet presents the feature selection evaluation and the third presents the evaluation of our model before the application of the chi-square test and after